Fix sorting bug on project monitor page
When sorting by a certain column the cells with the same content weren't
always placed together.

The problem is that Datatables sorts by the content of the cell, which
is not a simple value ('unresolvable', 'failed'...) but an HTML tag to
be displayed. So, it's trying to sort by values like "<a href=...".

Now we specify the value to be displayed and the value to sort/filter
by.

Co-authored-by: Henne Vogelsang <hvogel@opensuse.org>